requirements.txt
einops==0.4.0
fairscale==0.4.13
flash_attn==2.8.3
higher==0.2.1
hydra-core==1.3.2
iopath==0.1.10
matplotlib==3.5.1
nltk==3.6.5
numpy==2.3.5
omegaconf==2.3.0
openai==0.27.9
opencv_python==4.8.0.76
packaging==25.0
pandas==1.4.0
peft==0.9.0
Pillow==10.0.1
Pillow==12.0.0
PyYAML==6.0
PyYAML==6.0.3
Requests==2.32.5
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence_transformers==2.2.2
setuptools==68.0.0
timm==0.9.7
torch==2.0.1
torchvision==0.15.2
tqdm==4.62.3
transformers==4.40.0
